Retain previous test->source mapping on test failure
In the case of a test failure, the test might end before all
the code that needs to be exercised in the test gets a chance to run.
For this scenario: merge the test->source mapping from the previous run
to ensure that the next run triggers the test even if the test itself
doesn't change and only the unexercised paths change.